Fixing numpy update
