+++ service_funcs.py
@@ -1,0 +1,13 @@
+def timeit(method):
+    """Метод-декоратор для подсчета времени выполнения функции"""
+    def timed(*args, **kw):
+        ts = time.time()
+        result = method(*args, **kw)
+        te = time.time()
+        if 'log_time' in kw:
+            name = kw.get('log_name', method.__name__.upper())
+            kw['log_time'][name] = int((te - ts) * 1000)
+        else:
+            print(f'{method.__name__}  {(te - ts) * 1000}')
+        return result
+    return timed
